refactor: moving login image into a separate comp
